clara_5.2/requirements.txt
# Core runtime
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
pydantic==2.9.0
python-multipart==0.0.9
requests==2.32.3

# File & text processing
Pillow==10.4.0
pdf2image==1.17.0
PyPDF2==3.0.1
pdfminer.six==20240706
python-docx==1.1.2
langdetect==1.0.9
pandas==2.2.2

# OCR (requires Tesseract installed separately)
pytesseract==0.3.13

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
